FitnessFunction Updated
FitnessFunction class now calculates the fitness of an individual and saves it as an attribute of the individual. The fitness function is calculated using the torque required for moving the joint minus the gravity torque. The sum of all the torques' absolute values is then added to the cartesian distance of the final point to the desired final point, then a fitness function is calculated using both of these numbers.